+++ src/collision/Collider.h
@@ -18,7 +18,7 @@ enum class CollisionShape {
 };
 
 struct CollisionInfo {
-    float penetration = 0.0f;
+    double penetration = 0.0;
     math::Vec3 normal{};
 };
 
@@ -34,7 +34,7 @@ public:
     virtual bool testPoint(const math::Vec3& point) const = 0;
 
     // ray thickness
-    virtual float computeThickness(const math::Vec3& rayOrigin, const math::Vec3& rayDir) const { return 1e6f; }
+    virtual double computeThickness(const math::Vec3& rayOrigin, const math::Vec3& rayDir) const { return 1e6; }
 
     void setMaterial(const terminal::Material& mat) { m_material = mat; }
     const std::optional<terminal::Material>& getMaterial() const { return m_material; }
